+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.2.1
 
 - Separate background activity flags to avoid crosstalk between processes
+- Autoscaler::StubScaler may be used for local testing
 
 ## 0.2.0